fix(http): remove re-exports from hyper
