+++ src/http/mod.rs
@@ -1,15 +1,16 @@
 //! Low level HTTP definitions from Hyper
 
 mod body;
-mod into_response;
 mod segments;
-pub(crate) mod request;
+mod request;
 
-pub use hyper::{header, mime, Method, Request as HyperRequest, Response, StatusCode};
-pub use hyper::header::{Header, Headers};
-pub use http_crate::{Extensions, Request as HttpRequest, Response as HttpResponse};
+pub use http_crate::Extensions;
 
 pub use self::body::{Body, BodyStream, FromBody};
-pub use self::into_response::IntoResponse;
 pub use self::request::{Request, RequestParts};
 pub use self::segments::{FromSegment, FromSegments, Segment, Segments};
+
+#[allow(missing_docs)]
+pub trait FromHeader: ::hyper::header::Header + Clone {}
+
+impl<H: ::hyper::header::Header + Clone> FromHeader for H {}
